# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/requests": {
      "post": {
        "operationId": "RequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      },
      "get": {
        "operationId": "RequestsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      }
    },
    "/v1/requests/{id}": {
      "get": {
        "operationId": "RequestsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      },
      "patch": {
        "operationId": "RequestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      }
    },
...
    "/v1/company/requests": {
      "get": {
        "operationId": "CompanyQuotesController_listRequests",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/company/requests/{id}": {
      "get": {
        "operationId": "CompanyQuotesController_getRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/company/requests/{id}/quotes": {
      "post": {
        "operationId": "CompanyQuotesController_createQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/company/quotes": {
      "get": {
        "operationId": "CompanyQuotesController_listQuotes",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/company/quotes/{id}": {
      "patch": {
        "operationId": "CompanyQuotesController_updateQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/company/quotes/{id}/withdraw": {
      "post": {
        "operationId": "CompanyQuotesController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Firma: talepler ve teklifler"
        ]
      }
    },
    "/v1/requests/{id}/quotes": {
      "get": {
        "operationId": "CustomerQuotesController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      }
    },
    "/v1/quotes/{id}/accept": {
      "post": {
        "operationId": "CustomerQuotesController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Taşıma talepleri (müşteri)"
        ]
      }
    },
...
    "/v1/admin/companies": {
      "get": {
        "operationId": "AdminCompaniesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "VERIFIED",
                "REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin: firmalar"
        ]
      }
    },
    "/v1/admin/companies/{id}/verify": {
      "post": {
        "operationId": "AdminCompaniesController_verify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin: firmalar"
        ]
      }
    },
    "/v1/admin/companies/{id}/reject": {
      "post": {
        "operationId": "AdminCompaniesController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin: firmalar"
        ]
      }
    }
  },
  "info": {
    "title": "evdenevenakliyat.app API",
    "description": "Evden eve taşınma pazaryeri REST API. Korunan uç noktalar için `Authorization: Bearer <accessToken>` başlığı gerekir.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "fromCityCode": {
            "type": "string",
            "example": "34",
            "description": "Çıkış ili plaka kodu"
          },
          "fromDistrict": {
            "type": "string",
            "example": "kadikoy",
            "description": "Çıkış ilçesi adres kodu (GET /locations/cities/:code/districts)"
          },
          "fromAddress": {
            "type": "string",
            "example": "Caferağa Mah. Moda Cad. No:1 D:5"
          },
          "fromFloor": {
            "type": "number",
            "example": 3,
            "description": "Kat (bodrum için -1)"
          },
          "fromHasElevator": {
            "type": "boolean"
          },
          "toCityCode": {
            "type": "string",
            "example": "06"
          },
          "toDistrict": {
            "type": "string",
            "example": "cankaya"
          },
          "toAddress": {
            "type": "string",
            "example": "Kızılay Mah. Atatürk Blv. No:10 D:2"
          },
          "toFloor": {
            "type": "number",
            "example": 2
          },
          "toHasElevator": {
            "type": "boolean"
          },
          "homeType": {
            "type": "string",
            "enum": [
              "STUDIO",
              "ONE_PLUS_ONE",
              "TWO_PLUS_ONE",
              "THREE_PLUS_ONE",
              "FOUR_PLUS_ONE",
              "VILLA",
              "OFFICE"
            ],
            "example": "TWO_PLUS_ONE"
          },
          "moveDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-11-15",
            "description": "Taşınma tarihi (en erken yarın, en geç 1 yıl sonra)"
          },
          "isDateFlexible": {
            "type": "boolean",
            "default": false
          },
          "needsPacking": {
            "type": "boolean",
            "default": false
          },
          "needsAssembly": {
            "type": "boolean",
            "default": false
          },
          "needsStorage": {
            "type": "boolean",
            "default": false
          },
          "specialItems": {
            "example": [
              "Piyano",
              "Antika vitrin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Sokak dar, kamyon girmiyor."
          }
        },
        "required": [
          "fromCityCode",
          "fromDistrict",
          "fromAddress",
          "fromFloor",
          "fromHasElevator",
          "toCityCode",
          "toDistrict",
          "toAddress",
          "toFloor",
          "toHasElevator",
          "homeType",
          "moveDate"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "fromCityCode": {
            "type": "string",
            "example": "34",
            "description": "Çıkış ili plaka kodu"
          },
          "fromDistrict": {
            "type": "string",
            "example": "kadikoy",
            "description": "Çıkış ilçesi adres kodu (GET /locations/cities/:code/districts)"
          },
          "fromAddress": {
            "type": "string",
            "example": "Caferağa Mah. Moda Cad. No:1 D:5"
          },
          "fromFloor": {
            "type": "number",
            "example": 3,
            "description": "Kat (bodrum için -1)"
          },
          "fromHasElevator": {
            "type": "boolean"
          },
          "toCityCode": {
            "type": "string",
            "example": "06"
          },
          "toDistrict": {
            "type": "string",
            "example": "cankaya"
          },
          "toAddress": {
            "type": "string",
            "example": "Kızılay Mah. Atatürk Blv. No:10 D:2"
          },
          "toFloor": {
            "type": "number",
            "example": 2
          },
          "toHasElevator": {
            "type": "boolean"
          },
          "homeType": {
            "type": "string",
            "enum": [
              "STUDIO",
              "ONE_PLUS_ONE",
              "TWO_PLUS_ONE",
              "THREE_PLUS_ONE",
              "FOUR_PLUS_ONE",
              "VILLA",
              "OFFICE"
            ],
            "example": "TWO_PLUS_ONE"
          },
          "moveDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-11-15",
            "description": "Taşınma tarihi (en erken yarın, en geç 1 yıl sonra)"
          },
          "isDateFlexible": {
            "type": "boolean",
            "default": false
          },
          "needsPacking": {
            "type": "boolean",
            "default": false
          },
          "needsAssembly": {
            "type": "boolean",
            "default": false
          },
          "needsStorage": {
            "type": "boolean",
            "default": false
          },
          "specialItems": {
            "example": [
              "Piyano",
              "Antika vitrin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Sokak dar, kamyon girmiyor."
          }
        }
      },
...
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "priceTry": {
            "type": "number",
            "example": 18500,
            "description": "KDV dahil toplam fiyat (TL)"
          },
          "includesPacking": {
            "type": "boolean",
            "default": false
          },
          "includesAssembly": {
            "type": "boolean",
            "default": false
          },
          "includesInsurance": {
            "type": "boolean",
            "default": false
          },
          "crewSize": {
            "type": "number",
            "example": 3
          },
          "vehicleType": {
            "type": "string",
            "enum": [
              "PANELVAN",
              "KAMYONET",
              "KAMYON",
              "TIR"
            ],
            "example": "KAMYON"
          },
          "message": {
            "type": "string",
            "example": "Asansörlü taşıma dahildir."
          },
          "validUntil": {
            "format": "date-time",
            "type": "string",
            "description": "Teklifin geçerlilik sonu. Verilmezse talep süresi sonuna kadar geçerli."
          }
        },
        "required": [
          "priceTry",
          "crewSize",
          "vehicleType"
        ]
      },
      "UpdateQuoteDto": {
        "type": "object",
        "properties": {
          "priceTry": {
            "type": "number",
            "example": 18500,
            "description": "KDV dahil toplam fiyat (TL)"
          },
          "includesPacking": {
            "type": "boolean",
            "default": false
          },
          "includesAssembly": {
            "type": "boolean",
            "default": false
          },
          "includesInsurance": {
            "type": "boolean",
            "default": false
          },
          "crewSize": {
            "type": "number",
            "example": 3
          },
          "vehicleType": {
            "type": "string",
            "enum": [
              "PANELVAN",
              "KAMYONET",
              "KAMYON",
              "TIR"
            ],
            "example": "KAMYON"
          },
          "message": {
            "type": "string",
            "example": "Asansörlü taşıma dahildir."
          },
          "validUntil": {
            "format": "date-time",
            "type": "string",
            "description": "Teklifin geçerlilik sonu. Verilmezse talep süresi sonuna kadar geçerli."
          }
        }
      },
      "RejectCompanyDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "K3 belgesinin süresi dolmuş, güncel belgeyi yükleyin."
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}